// src/app/models/pedido.ts
export class Pedido {

    cgcCliente: string;
    idTransportadora: number;
    codigoCondicaoPagamento: number;
    codigoVendedor: number;
    codigoTabela: number;
    codigoDesconto: number;
    numeroOC: string;
    dataEntrega: string;
    dataBaseVencimento: string;
    valorFrete: number;
    valorDesconto: number;
    observacoes: string;

}